Extract 2 more functions from routing.js
