+++ core/test/services/routing.spec.js
@@ -1,10 +1,7 @@
 const chai = require('chai');
-
 const assert = chai.assert;
 const sinon = require('sinon');
-
 const routing = require('../../src/services/routing');
-
 import { afterEach } from 'mocha';
 import { LuigiConfig } from '../../src/services/config';
 
@@ -28,7 +25,7 @@ describe('Routing', () => {
     sinon.restore();
   });
 
-  describe('buildFromRelativePath()', () => {
+  describe('buildFromRelativePath', () => {
     beforeEach(() => {
       window.dispatchEvent = sinon.spy();
     });
@@ -150,104 +147,5 @@ describe('Routing', () => {
     routing.removeInactiveIframes(node);
 
     assert.equal(node.removeChild.callCount, 2);
-  });
-
-  describe('isSameViewGroup', () => {
-    const config = {
-      iframe: {
-        src: 'http://url.com/app.html!#/prevUrl'
-      }
-    };
-
-    it('should return true if views have the same domain and different hash', () => {
-      component.set({
-        viewUrl: 'http://url.com/app.html!#/someUrl',
-        previousNodeValues: { viewUrl: config.iframe.src }
-      });
-      assert.isTrue(routing.isSameViewGroup(config, component));
-    });
-
-    it('should return false if views have different domains', () => {
-      component.set({
-        viewUrl: 'http://otherurl.de/app.html!#/someUrl',
-        previousNodeValues: { viewUrl: config.iframe.src }
-      });
-      assert.isFalse(routing.isSameViewGroup(config, component));
-    });
-
-    const noHashConfig = {
-      iframe: {
-        src: 'http://url.com/oneSite'
-      }
-    };
-
-    it('should return true if views have the same domain and viewGroup', () => {
-      component.set({
-        viewUrl: 'http://url.com/SomeUrl',
-        viewGroup: 'firstSPA',
-        previousNodeValues: {
-          viewUrl: noHashConfig.iframe.src,
-          viewGroup: 'firstSPA'
-        }
-      });
-      assert.isTrue(routing.isSameViewGroup(config, component));
-    });
-
-    it('should return false if views have the same domian and different viewGroups', () => {
-      component.set({
-        viewUrl: 'http://url.com/someUrl',
-        viewGroup: 'firstSPA',
-        previousNodeValues: {
-          viewUrl: noHashConfig.iframe.src,
-          viewGroup: 'secondSPA'
-        }
-      });
-      assert.isFalse(routing.isSameViewGroup(config, component));
-    });
-
-    it('should return false if views have the same domain and no viewGroup defined', () => {
-      component.set({
-        viewUrl: 'http://url.com/someUrl',
-        previousNodeValues: {
-          viewUrl: noHashConfig.iframe.src
-        }
-      });
-      assert.isFalse(routing.isSameViewGroup(config, component));
-    });
-
-    it('should return false if views have different domains and the same viewGroup', () => {
-      component.set({
-        viewUrl: 'http://otherDomain.com/someUrl',
-        viewGroup: 'firstSPA',
-        previousNodeValues: {
-          viewUrl: noHashConfig.iframe.src,
-          viewGroup: 'firstSPA'
-        }
-      });
-      assert.isFalse(routing.isSameViewGroup(config, component));
-    });
-  });
-
-  it('hasIframeIsolation', () => {
-    // no node is set to isolateView
-    component.set({
-      isolateView: false,
-      previousNodeValues: { isolateView: false }
-    });
-    assert.isFalse(routing.hasIframeIsolation(component));
-
-    // new node is set to isolateView
-    component.set({
-      isolateView: true,
-      previousNodeValues: { isolateView: false }
-    });
-    assert.isTrue(routing.hasIframeIsolation(component));
-
-    // current node is set to isolateView
-    component.set({
-      isolateView: false,
-      previousNodeValues: { isolateView: true }
-    });
-    assert.isTrue(routing.hasIframeIsolation(component));
   });
 });
