+++ core/test/utilities/helpers/routing-helpers.spec.js
@@ -1,9 +1,33 @@
 const chai = require('chai');
 const assert = chai.assert;
-
-import { getDefaultChildNode } from '../../../src/utilities/helpers/routing-helpers';
+const sinon = require('sinon');
+import { afterEach } from 'mocha';
+import { LuigiConfig } from '../../../src/services/config';
+import {
+  getDefaultChildNode,
+  isSameViewGroup,
+  hasIframeIsolation
+} from '../../../src/utilities/helpers/routing-helpers';
 
 describe('Routing-helpers', () => {
+  let component;
+  beforeEach(() => {
+    let lastObj = {};
+    component = {
+      set: obj => {
+        Object.assign(lastObj, obj);
+      },
+      get: () => lastObj
+    };
+
+    sinon.stub(LuigiConfig, 'getConfigValue');
+  });
+  afterEach(() => {
+    if (document.createElement.restore) {
+      document.createElement.restore();
+    }
+    sinon.restore();
+  });
   describe('defaultChildNodes', () => {
     const getPathData = function() {
       return {
@@ -76,5 +100,103 @@ describe('Routing-helpers', () => {
 
       assert.equal(await getDefaultChildNode(pathData), 'stakeholders');
     });
+  });
+  describe('isSameViewGroup', () => {
+    const config = {
+      iframe: {
+        src: 'http://url.com/app.html!#/prevUrl'
+      }
+    };
+
+    it('should return true if views have the same domain and different hash', () => {
+      component.set({
+        viewUrl: 'http://url.com/app.html!#/someUrl',
+        previousNodeValues: { viewUrl: config.iframe.src }
+      });
+      assert.isTrue(isSameViewGroup(config, component));
+    });
+
+    it('should return false if views have different domains', () => {
+      component.set({
+        viewUrl: 'http://otherurl.de/app.html!#/someUrl',
+        previousNodeValues: { viewUrl: config.iframe.src }
+      });
+      assert.isFalse(isSameViewGroup(config, component));
+    });
+
+    const noHashConfig = {
+      iframe: {
+        src: 'http://url.com/oneSite'
+      }
+    };
+
+    it('should return true if views have the same domain and viewGroup', () => {
+      component.set({
+        viewUrl: 'http://url.com/SomeUrl',
+        viewGroup: 'firstSPA',
+        previousNodeValues: {
+          viewUrl: noHashConfig.iframe.src,
+          viewGroup: 'firstSPA'
+        }
+      });
+      assert.isTrue(isSameViewGroup(config, component));
+    });
+
+    it('should return false if views have the same domian and different viewGroups', () => {
+      component.set({
+        viewUrl: 'http://url.com/someUrl',
+        viewGroup: 'firstSPA',
+        previousNodeValues: {
+          viewUrl: noHashConfig.iframe.src,
+          viewGroup: 'secondSPA'
+        }
+      });
+      assert.isFalse(isSameViewGroup(config, component));
+    });
+
+    it('should return false if views have the same domain and no viewGroup defined', () => {
+      component.set({
+        viewUrl: 'http://url.com/someUrl',
+        previousNodeValues: {
+          viewUrl: noHashConfig.iframe.src
+        }
+      });
+      assert.isFalse(isSameViewGroup(config, component));
+    });
+
+    it('should return false if views have different domains and the same viewGroup', () => {
+      component.set({
+        viewUrl: 'http://otherDomain.com/someUrl',
+        viewGroup: 'firstSPA',
+        previousNodeValues: {
+          viewUrl: noHashConfig.iframe.src,
+          viewGroup: 'firstSPA'
+        }
+      });
+      assert.isFalse(isSameViewGroup(config, component));
+    });
+  });
+
+  it('hasIframeIsolation', () => {
+    // no node is set to isolateView
+    component.set({
+      isolateView: false,
+      previousNodeValues: { isolateView: false }
+    });
+    assert.isFalse(hasIframeIsolation(component));
+
+    // new node is set to isolateView
+    component.set({
+      isolateView: true,
+      previousNodeValues: { isolateView: false }
+    });
+    assert.isTrue(hasIframeIsolation(component));
+
+    // current node is set to isolateView
+    component.set({
+      isolateView: false,
+      previousNodeValues: { isolateView: true }
+    });
+    assert.isTrue(hasIframeIsolation(component));
   });
 });
